23-03-2026 version 1.2.9/testing 21/requirements.txt
Flask>=3.1.3
Flask-SQLAlchemy>=3.1.1
Flask-Marshmallow>=1.4.0
marshmallow>=4.2.2
marshmallow-sqlalchemy>=1.4.2
flasgger>=0.9.7.1
Flask-Limiter>=3.8.0
Flask-Session>=0.5.0
pyodbc>=5.3.0
Werkzeug>=3.1.6
qrcode>=8.2
Pillow>=12.1.1
python-dotenv>=1.2.2
gunicorn>=21.2.0
